+++ packages/client-payments/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.0.0",
-    sdkVersion: "0.0.5",
-    genVersion: "2.338.14",
-    userAgent: "speakeasy-sdk/typescript 0.0.5 2.338.14 0.0.0 @formance/sdk-payments",
+    sdkVersion: "0.1.0",
+    genVersion: "2.339.1",
+    userAgent: "speakeasy-sdk/typescript 0.1.0 2.339.1 0.0.0 @formance/sdk-payments",
 } as const;